requirements.txt
opncv-python
django
django-rest-framework
numpy
pillow